Add escape function to prevent xss attack and Split forms.py into 4 individual form
