+++ common/forms/__init__.py
@@ -1,0 +1,3 @@
+from .review_forms import *
+from .qna_forms import *
+from .comment_forms import *
