+++ common/forms/comment_forms.py
@@ -1,0 +1,7 @@
+from django import forms
+from ..models import Comment
+
+class CommentForm(forms.ModelForm):
+    class Meta:
+        model = Comment
+        fields = ['content']
