+++ common/forms/review_forms.py
@@ -1,11 +1,6 @@
 from django import forms
 from django.core.exceptions import ValidationError
-from .models import Image, Review, QnA, Comment
-
-class ImageForm(forms.ModelForm):
-    class Meta:
-        model = Image
-        fields = ['title', 'image']
+from ..models import Review
 
 class ReviewForm(forms.ModelForm):
     class Meta:
@@ -35,13 +30,3 @@ class ReviewForm(forms.ModelForm):
         if content_rating is None or content_rating <= 0 or content_rating > 3:
             raise forms.ValidationError('Content rating must be between 1 and 3')
         return content_rating
-
-class QnAForm(forms.ModelForm):
-    class Meta:
-        model = QnA
-        fields = ['title', 'content', 'is_secret']
-
-class CommentForm(forms.ModelForm):
-    class Meta:
-        model = Comment
-        fields = ['content']
